Move Analytics.tsx to Utils
